+++ R/rwtbl_helpers.R
@@ -4,7 +4,7 @@
 #' `rwtbl_slot_names()` lists all of the slot names found in a `tbl_df` object 
 #' containing RiverWare output data.
 #' 
-#' Given a `tbl_df` object that is returned by [rw_rdf_to_tbl()] or 
+#' Given a `tbl_df` object that is returned by [rdf_to_rwtbl()] or 
 #' [rw_read_rw_csv()], return all of the Object.Slot names found in the data 
 #' frame. These are the unique full slot names found in the `ObjectSlot` 
 #' column.
@@ -12,10 +12,10 @@
 #' @param rwtbl A `tbl_df` object with RiverWare output. Must contain the 
 #'   `ObjectSlot` column.
 #' 
-#' @seealso [rw_rdf_to_tbl()], [rw_read_rw_csv()]
+#' @seealso [rdf_to_rwtbl()], [rw_read_rw_csv()]
 #' 
 #' @examples 
-#' rwtbl <- rw_rdf_to_tbl(keyRdf)
+#' rwtbl <- rdf_to_rwtbl(keyRdf)
 #' rwtbl_slot_names(rwtbl)
 #' 
 #' @export