+++ tests/testthat/test_apply_eval.R
@@ -6,7 +6,7 @@ slot_agg_matrix <- rwd_agg(read.csv(
   stringsAsFactors = FALSE
 ))[1:5,]
 
-rwtbl <- rw_rdf_to_tbl(keyRdf)
+rwtbl <- rdf_to_rwtbl(keyRdf)
 
 # check results --------------------------
 test_that("apply_eval works as expected", {