+++ tests/testthat/test_rdf_to_tbl.R
@@ -1,9 +1,9 @@
 library(RWDataPlyr)
 
-context("check rw_rdf_to_tbl()")
+context("check rdf_to_rwtbl()")
 
 # check the default call -----------------
-rdftbl <- rw_rdf_to_tbl(keyRdf)
+rdftbl <- rdf_to_rwtbl(keyRdf)
 reqCols <- RWDataPlyr:::req_rwtbl_cols()
 exp_atts <- c("mrm_config_name", "owner", "description", "create_date", 
               "n_traces")
@@ -42,7 +42,7 @@ test_that("no values are modified", {
 })
 
 # check the add_ym options ---------------
-rdftbl2 <- rw_rdf_to_tbl(keyRdf, add_ym = FALSE)
+rdftbl2 <- rdf_to_rwtbl(keyRdf, add_ym = FALSE)
 test_that("dimensions are as expected with no ym", {
   expect_true(nrow(rdftbl2) > 0)
   expect_equal(ncol(rdftbl2), length(reqCols))
@@ -54,15 +54,15 @@ test_that("no ym tbl matches orig tbl", {
 })
 
 test_that("invalid add_ym values cause errors", {
-  expect_error(rw_rdf_to_tbl(keyRdf, add_ym = NA))
-  expect_error(rw_rdf_to_tbl(keyRdf, add_ym = "true"))
-  expect_error(rw_rdf_to_tbl(keyRdf, add_ym = 7))
-  expect_error(rw_rdf_to_tbl(keyRdf, add_ym = c(FALSE, TRUE, NA)))
+  expect_error(rdf_to_rwtbl(keyRdf, add_ym = NA))
+  expect_error(rdf_to_rwtbl(keyRdf, add_ym = "true"))
+  expect_error(rdf_to_rwtbl(keyRdf, add_ym = 7))
+  expect_error(rdf_to_rwtbl(keyRdf, add_ym = c(FALSE, TRUE, NA)))
 })
 
 # check the scenario option -------------
-rdftbl2 <- rw_rdf_to_tbl(keyRdf, scenario = "DNF,CT")
-rdftbl3 <- rw_rdf_to_tbl(keyRdf, scenario = 1, add_ym = FALSE)
+rdftbl2 <- rdf_to_rwtbl(keyRdf, scenario = "DNF,CT")
+rdftbl3 <- rdf_to_rwtbl(keyRdf, scenario = 1, add_ym = FALSE)
 test_that("dimensions are as expected with no ym", {
   expect_true(nrow(rdftbl2) > 0)
   expect_equal(ncol(rdftbl2), length(reqCols) + 3)
@@ -87,14 +87,14 @@ test_that("no add scenario tbl matches orig tbl", {
 })
 
 test_that("scenario options error properly", {
-  expect_error(rw_rdf_to_tbl(keyRdf, scenario = c("DNF,CT", "DNF,C1")))
-  expect_error(rw_rdf_to_tbl(keyRdf, scenario = 1:6, add_ym = FALSE))
-  expect_error(rw_rdf_to_tbl(keyRdf, scenario = character(0), add_ym = FALSE))
+  expect_error(rdf_to_rwtbl(keyRdf, scenario = c("DNF,CT", "DNF,C1")))
+  expect_error(rdf_to_rwtbl(keyRdf, scenario = 1:6, add_ym = FALSE))
+  expect_error(rdf_to_rwtbl(keyRdf, scenario = character(0), add_ym = FALSE))
 })
 
 # check the keep_cols option ------------
-rdftbl2 <- rw_rdf_to_tbl(keyRdf, keep_cols = TRUE)
-rdftbl3 <- rw_rdf_to_tbl(keyRdf, keep_cols = c("ObjectName", "Unit"), scenario = 1)
+rdftbl2 <- rdf_to_rwtbl(keyRdf, keep_cols = TRUE)
+rdftbl3 <- rdf_to_rwtbl(keyRdf, keep_cols = c("ObjectName", "Unit"), scenario = 1)
 test_that("dimensions are as expected with keep_cols", {
   expect_true(nrow(rdftbl2) > 0)
   expect_equal(ncol(rdftbl2), length(reqCols) + 8)
@@ -114,7 +114,7 @@ test_that("dimensions are as expected with keep_cols", {
 
 test_that("keep_cols warnings post correctly", {
   expect_warning(
-    tmp <- rw_rdf_to_tbl(keyRdf, keep_cols = c("ObjectName", "missing")),
+    tmp <- rdf_to_rwtbl(keyRdf, keep_cols = c("ObjectName", "missing")),
     paste0(
       "The following columns specified by 'keep_cols' were not found in the rwtbl:\n",
       "    missing"
@@ -124,7 +124,7 @@ test_that("keep_cols warnings post correctly", {
   expect_true(all(colnames(tmp) %in% c(reqCols, "Year", "Month", "ObjectName")))
   
   expect_warning(
-    tmp <- rw_rdf_to_tbl(keyRdf, keep_cols = c("ObjectName", "missing", "a")),
+    tmp <- rdf_to_rwtbl(keyRdf, keep_cols = c("ObjectName", "missing", "a")),
     paste0(
       "The following columns specified by 'keep_cols' were not found in the rwtbl:\n",
       "    missing, a"
@@ -135,9 +135,9 @@ test_that("keep_cols warnings post correctly", {
 
 
 # check the default call annual values-----------------
-context("check rw_rdf_to_tbl with annual rdf file")
+context("check rdf_to_rwtbl with annual rdf file")
 
-rdftbl <- rw_rdf_to_tbl(sysRdf)
+rdftbl <- rdf_to_rwtbl(sysRdf)
 reqCols <- RWDataPlyr:::req_rwtbl_cols()
 nt <- as.integer(sysRdf$meta$number_of_runs)
 slots <- rdf_slot_names(sysRdf)
@@ -174,7 +174,7 @@ test_that("values are as expected", {
   expect_true(all(rdftbl$Month == "December"))
 })
 
-context("check rw_rdf_to_tbl() internal helpers")
+context("check rdf_to_rwtbl() internal helpers")
 
 test_that("req_rwtbl_cols()", {
   expect_type(RWDataPlyr:::req_rwtbl_cols(), "character") %>%